Move a treebank modifying filter to the treebank directory, why not
