+++ src/edu/stanford/nlp/trees/treebank/DuplicateTreeStringFilter.java
@@ -1,4 +1,6 @@
-package edu.stanford.nlp.trees;
+package edu.stanford.nlp.trees.treebank;
+
+import edu.stanford.nlp.trees.Tree;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -8,7 +10,7 @@ import java.util.function.Predicate;
  * Filters trees based on duplicate toString()
  * <br>
  * for example, 
- * <code>java edu.stanford.nlp.trees.Treebanks -filter edu.stanford.nlp.trees.DuplicateTreeStringFilter -pennPrint /u/nlp/data/constituency-parser/models-4.0.0/data/ewt/ptb/train/ewt-train.mrg</code>
+ * <code>java edu.stanford.nlp.trees.Treebanks -filter edu.stanford.nlp.trees.treebank.DuplicateTreeStringFilter -pennPrint /u/nlp/data/constituency-parser/models-4.0.0/data/ewt/ptb/train/ewt-train.mrg</code>
  *
  * @author John Bauer
  */
